Engine calculate pay as without promotions.

+++ src/main/java/com/ynz/enginee/promotion/enginee/Payable.java
@@ -1,5 +1,5 @@
 package com.ynz.enginee.promotion.enginee;
 
 public interface Payable {
-    Double calculatePayment();
+    Integer calculatePayment();
 }
