+++ src/main/java/com/ynz/enginee/promotion/enginee/promotions/BuyNItems.java
@@ -10,7 +10,8 @@ public class BuyNItems extends AbstractPromotion {
     }
 
     @Override
-    public Double calculatePayment() {
+    public Integer calculatePayment() {
+
         return super.calculatePayment();
     }
 }
